% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello50-package.R
\name{hello50}
\alias{hello50}
\title{It does some stuff}
\description{
It does some stuff
}
\section{Support}{


The development repository for the **hello50** package is found at
\url{https://github.com/noamross/hello50}.  Please file
bug reports or other feedback at \url{https://github.com/noamross/hello50/issues}
}

\author{
Noam Ross \email{noam.ross@gmail.com}
}
\keyword{package}
